+++ psutil/_pslinux.py
@@ -47,7 +47,6 @@ from ._common import path_exists_strict
 from ._common import supports_ipv6
 from ._common import usage_percent
 
-
 # fmt: off
 __extra__all__ = [
     'PROCFS_PATH',
@@ -2021,7 +2020,7 @@ class Process:
                     else:
                         try:
                             data[fields[0]] = int(fields[1]) * 1024
-                        except ValueError:
+                        except (ValueError, IndexError):
                             if fields[0].startswith(b'VmFlags:'):
                                 # see issue #369
                                 continue