+++ psutil/arch/freebsd/proc.c
@@ -349,7 +349,6 @@ error:
 }
 
 
-#if defined(__FreeBSD_version) && __FreeBSD_version >= 701000
 PyObject *
 psutil_proc_cwd(PyObject *self, PyObject *args) {
     pid_t pid;
@@ -397,10 +396,8 @@ error:
         free(freep);
     return NULL;
 }
-#endif
 
 
-#if defined(__FreeBSD_version) && __FreeBSD_version >= 800000
 PyObject *
 psutil_proc_num_fds(PyObject *self, PyObject *args) {
     pid_t pid;
@@ -424,7 +421,6 @@ psutil_proc_num_fds(PyObject *self, PyObject *args) {
 
     return Py_BuildValue("i", cnt);
 }
-#endif
 
 
 PyObject *