+++ psutil/arch/sunos/environ.c
@@ -10,17 +10,13 @@
 
 #include <Python.h>
 
-#if !defined(_LP64) && _FILE_OFFSET_BITS == 64
-    #undef _FILE_OFFSET_BITS
-    #undef _LARGEFILE64_SOURCE
-#endif
-
 #include <sys/types.h>
 #include <sys/procfs.h>
 #include <sys/stat.h>
 #include <fcntl.h>
 #include <unistd.h>
 
+#include "../../arch/all/init.h"
 #include "environ.h"