+++ psutil/tests/test_posix.py
@@ -34,7 +34,6 @@ from psutil.tests import skip_on_access_denied
 from psutil.tests import spawn_subproc
 from psutil.tests import terminate
 
-
 if POSIX:
     import mmap
     import resource