+++ psutil/tests/test_process.py
@@ -64,7 +64,6 @@ from psutil.tests import skip_on_access_denied
 from psutil.tests import skip_on_not_implemented
 from psutil.tests import wait_for_pid
 
-
 # ===================================================================
 # --- psutil.Process class tests
 # ===================================================================
@@ -748,15 +747,20 @@ class TestProcess(PsutilTestCase):
             ["-c", "import time; [time.sleep(0.1) for x in range(100)]"]
         )
         p = self.spawn_psproc(cmdline)
+
+        # XXX - flaky test: exclude the python exe which, for some
+        # reason, and only sometimes, on OSX appears different.
+        cmdline = cmdline[1:]
+
         if OPENBSD:
             # XXX: for some reason the test process may turn into a
             # zombie (don't know why).
             try:
-                assert p.cmdline() == cmdline
+                assert p.cmdline()[1:] == cmdline
             except psutil.ZombieProcess:
                 raise pytest.skip("OPENBSD: process turned into zombie")
         else:
-            ret = p.cmdline()
+            ret = p.cmdline()[1:]
             if NETBSD and ret == []:
                 # https://github.com/giampaolo/psutil/issues/2250
                 raise pytest.skip("OPENBSD: returned EBUSY")
@@ -1519,9 +1523,9 @@ class TestProcess(PsutilTestCase):
             for name in exclude:
                 d.pop(name, None)
             return {
-                k.replace("\r", "").replace("\n", ""): v.replace(
-                    "\r", ""
-                ).replace("\n", "")
+                k.replace("\r", "").replace("\n", ""): (
+                    v.replace("\r", "").replace("\n", "")
+                )
                 for k, v in d.items()
             }