+++ scripts/internal/print_downloads.py
@@ -22,7 +22,6 @@ import pypinfo  # noqa: F401
 
 from psutil._common import memoize
 
-
 AUTH_FILE = os.path.expanduser("~/.pypinfo.json")
 PKGNAME = 'psutil'
 DAYS = 30