+++ scripts/internal/purge_installation.py
@@ -14,7 +14,6 @@ import os
 import shutil
 import site
 
-
 PKGNAME = "psutil"